Added a fcntl syscall (for setting socket file descriptors as non-blocking) and enabled non-blocking socket IO
Added a flag field to the file descriptor struct for things such as
the non-blocking flag.  This could be outside the kernel but for simplicity
sake (and for the fact that the file descriptor stuff should be rewritten soon)
it is easiest to add it to the kernel fd struct.