Added a basic seek syscall for files Doesn't support making holes right now (which cannot be made anyway since we cannot write to files.
Also fixed an alignment but that occured because the idle task tss's
are statically allocated but the mmx arena for the tss must be aligned
to a 16 byte address.  Added an alignment to the statically allocated tss
and a compiler assert so that the tss size is always a multiple of 32.