Fixed bugs related to ramdisk filesystem, it has now been tested
Note: The version of grub we are using for quest.iso does not seem to
like loading large modules and therefore if we dump everything we have
into the ramdisk it will not load.  I will look into updating the
version of grub we use to make quest.iso later.