Made changes to video and keyboard drivers and added syscalls for time and getting keyboard events
The enable_video syscall now checks the flag and if it isn't told to
set the video mode to use the color format used by qcv it just returns
the address the memory has been mapped to.

Added a syscall pressed_keys which returns which keys are pressed and
get_keyboard_events which returns the raw keyboard events.  These will
be useful for MAME.

Also implemented the times syscall.  It does not function properly in
the sense that it returns the value of a system wide tick and not how
long the process has been alive.  It also associates all time the user
space.  It is basically good enough to use the clock function to get a
time diff (which is needed by MAME).